megafile restructure
